+++ lib/contentful/support.rb
@@ -7,14 +7,6 @@ module Contentful
         snake = String(object).gsub(/(?<!^)[A-Z]/) { "_#$&" }
         snake.downcase
       end
-
-      # Transforms each hash key into a symbol (like in AS)
-      def symbolize_keys(hash)
-        result = {}
-        # XXX remove inline rescue
-        hash.each_key { |key| result[(key.to_sym rescue key)] = hash[key] }
-        result
-      end
     end
   end
 end