+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+### 0.2.0
+
+* Introduce new :entry_mapping configuration to enable custom Entry classes based on ContentTypes
+* Update HTTP gem dependency to 0.6
+* Convert arrays in query values to strings, separated by comma
+
+
 ### 0.1.3
 
 * Better link inclusion processing, prevent "stack level to deep" errors